refactor: moved all CLI related codes to cbsurge/cli folder
