+++ cbsurge/cli/__init__.py
@@ -1,11 +1,14 @@
-import logging
-
 from cbsurge.util.setup_logger import setup_logger
-from cbsurge.admin import admin
-from cbsurge.project import create, list, upload, download, publish, delete
-from cbsurge.initialize import init
-from cbsurge.assess import assess
-from cbsurge.az.auth import authenticate
+from cbsurge.cli.admin import admin
+from cbsurge.cli.auth import authenticate
+from cbsurge.cli.init import init
+from cbsurge.cli.assess import assess
+from cbsurge.cli.create import create
+from cbsurge.cli.delete import delete
+from cbsurge.cli.list import list
+from cbsurge.cli.upload import upload
+from cbsurge.cli.download import download
+from cbsurge.cli.publish import publish
 import click
