+++ cbsurge/session.py
@@ -1,12 +1,8 @@
 import logging
 import os
 import json
-
-from azure.identity import DefaultAzureCredential, AzureAuthorityHosts
 from azure.core.exceptions import ClientAuthenticationError
 from azure.storage.blob.aio import BlobServiceClient, ContainerClient
-from azure.storage.fileshare.aio import ShareServiceClient
-from cbsurge.az.auth import MsalTokenCredential, TOKEN_CACHE_FIlE
 from cbsurge.az.surgeauth import SurgeTokenCredential
 
 logger = logging.getLogger(__name__)
@@ -159,17 +155,6 @@ class Session(object):
             Azure TokenCredential is returned if authenticated.
         """
         credential = SurgeTokenCredential()
-        # # check if token file exists at user home folder
-        # config_path = os.path.dirname(self.get_config_file_path())
-        # token_cache_path = os.path.join(config_path, TOKEN_CACHE_FIlE)
-        # if os.path.exists(token_cache_path):
-        #     credential = MsalTokenCredential(device_auth=False)
-        # else:
-        #     # otherwise, use DefaultAzureCredential
-        #     credential = DefaultAzureCredential(
-        #         exclude_interactive_browser_credential=(not interactive_browser)
-        #     )
-
         return credential
 
 
@@ -198,11 +183,7 @@ class Session(object):
 
     def authenticate(self, scopes = "https://storage.azure.com/.default"):
         """
-        Authenticate to Azure through interactive browser if DefaultAzureCredential is not provideds.
-        Authentication uses DefaultAzureCredential.
-
-        Please refer to https://learn.microsoft.com/en-us/python/api/azure-identity/azure.identity.defaultazurecredential?view=azure-python
-        about DefaultAzureCredential api specificaiton.
+        Authenticate to Azure
 
         Parameters:
             scopes: scopes for get_token method. Default to "https://storage.azure.com/.default"
